rename filtermenu to mobilefiltermenu
